Ensure that php.d is not scanned when PHPRC is set
Case EA-4663: Since EasyApache 4 is built with the configuration
option 'with-config-file-scan-dir', this ensured that php.d was
read, allowing module configurations to be loaded. It resulted
in customer configurations not being overridden with any
situations where PHPRC is used (primarily suPHP). This patch
provides EasyApache 3 compatible behaviors when using
suPHP_ConfigPath, specifically by not scanning php.d when
PHPRC is set.